requirements.txt
pyspark==3.5.0
delta-spark==3.2.0
kafka-python
scikit-learn
faker